+++ graphlib/Graphs/Edge.cs
@@ -8,10 +8,10 @@ namespace graphlib
         /// <value> source is starting vertex of edge
         /// <value> destination is ending vertex of edge
         /// <value> weight is value or cost of edge
-        public long source {get; set;}
-        public long destination {get;set;}
+        public int source {get; set;}
+        public int destination {get;set;}
         public long weight {get;set;}
-        public Edge(long source, long destination, long weight)
+        public Edge(int source, int destination, long weight)
         {
             this.source = source;
             this.destination = destination;
